Generalise Language and Personal statement components
This makes it easier to use them in the support interface.
