+++ app/components/personal_statement_component.rb
@@ -1,0 +1,38 @@
+class PersonalStatementComponent < ActionView::Component::Base
+  validates :application_form, presence: true
+
+  delegate :becoming_a_teacher,
+           :subject_knowledge,
+           :interview_preferences,
+           :further_information,
+           to: :application_form
+
+  def initialize(application_form:)
+    @application_form = application_form
+  end
+
+  def rows
+    [
+      {
+        key: t('application_form.personal_statement.becoming_a_teacher.label'),
+        value: becoming_a_teacher,
+      },
+      {
+        key: 'Subject knowledge',
+        value: subject_knowledge,
+      },
+      {
+        key: t('application_form.personal_statement.interview_preferences.label'),
+        value: interview_preferences.presence || 'No preference given',
+      },
+      {
+        key: 'Further information',
+        value: further_information.presence || 'No further information given',
+      },
+    ]
+  end
+
+private
+
+  attr_reader :application_form
+end
