+++ app/views/provider_interface/application_choices/show.html.erb
@@ -46,11 +46,11 @@
 
 <h2 class="govuk-heading-l govuk-!-margin-top-8" id="language-skills">Language skills</h2>
 
-<%= render ProviderInterface::LanguageSkillsComponent, application_form: @application_choice.application_form %>
+<%= render LanguageSkillsComponent, application_form: @application_choice.application_form %>
 
 <h2 class="govuk-heading-l govuk-!-margin-top-8" id="personal-statement">Personal statement</h2>
 
-<%= render ProviderInterface::PersonalStatementComponent, application_form: @application_choice.application_form %>
+<%= render PersonalStatementComponent, application_form: @application_choice.application_form %>
 
 <h2 class="govuk-heading-l govuk-!-margin-top-8" id="first-reference">First reference</h2>
